Catch and warn of possible CVE-2011-0411 attacks
This class of attacks is possible when an existing connection is
upgraded to use TLS after being established.  A malicious attacker could
craft a man-in-the-middle attack wherein the "STARTTLS" command sent by
the client has additional malicious application-level commands appended
in the same packet.  The client and server are then left to complete the
TLS handshake on their own.

If the server does not clear the buffer during TLS authentication, it
will assume that the remaining commands in the buffer (inserted by the
attacker before TLS was established) were sent _after_ TLS was
established, and possibly run them with elevated privileges.  This
attack is particularly dangerous if the protocol makes use of SSL client
certificates to authenticate the client.

To prevent this, explicitly clear the handle's buffer while TLS is being
negotiated.